fix: change ply-loader color array type to Float32Array

+++ modules/ply/src/lib/normalize-ply.js
@@ -35,7 +35,7 @@ function normalizeAttributes(attributes) {
   }
 
   if (attributes.colors.length > 0) {
-    accessors.COLOR_0 = {value: new Uint8Array(attributes.colors), size: 3};
+    accessors.COLOR_0 = {value: new Float32Array(attributes.colors), size: 3};
   }
 
   return accessors;
